[BE,FE - HW-5] - Added Tags & Comments / added search stack to home page

+++ api/core/services/articles.service.ts
@@ -3,7 +3,7 @@ import { ArticleModel } from "../models/article.model";
 import { isEmpty } from "lodash";
 import { InternalError } from "../common/error-handler";
 import { UserJwtPayload } from "../dto/user.dto";
-import { Query } from "mongoose";
+import { PopulateOptions, Query } from "mongoose";
 import QueryString from "qs";
 import { QueryParamsParser } from "../common/query-params-parser";
 import { DatabaseQueryBuilder } from "../common/database.query-builder";
@@ -12,14 +12,21 @@ import TagsService from "./tags.service";
 import { isAdmin } from "../common/utils";
 
 class ArticlesService {
-  private static populateArticleQuery(query: Query<any, any>) {
-    return query
-      .populate({
+  private static buildArticlePopulateOptions(): PopulateOptions[] {
+    return [
+      {
         path: "comments",
         populate: { path: "author", select: { password: 0 } }
-      })
-      .populate("author", { password: 0 })
-      .populate("tags");
+      },
+      {
+        path: "author",
+        select: { password: 0 }
+      }
+    ];
+  }
+
+  private static populateArticleQuery(query: Query<any, any>) {
+    return query.populate(ArticlesService.buildArticlePopulateOptions());
   }
 
   async getAll(queryParams: QueryString.ParsedQs): Promise<ArticleDTO[]> {
@@ -33,26 +40,60 @@ class ArticlesService {
     const sortOptions = QueryParamsParser.parseSortParam(queryParams) || defaultSortOptions;
     query = DatabaseQueryBuilder.buildSortQuery(query, sortOptions);
     query = DatabaseQueryBuilder.setQueryIntersection(query, QueryParamsParser.parse(queryParams, "tags"));
-    query = DatabaseQueryBuilder.setQueryPagination(query, QueryParamsParser.parse(queryParams, "skip"));
-    query = DatabaseQueryBuilder.setQueryPagination(query, QueryParamsParser.parse(queryParams, "limit"));
+    // query = DatabaseQueryBuilder.setQueryPagination(query, QueryParamsParser.parse(queryParams, "skip"));
+    // query = DatabaseQueryBuilder.setQueryPagination(query, QueryParamsParser.parse(queryParams, "limit"));
 
     return ArticlesService.populateArticleQuery(query);
   }
 
   async getOne(_id: string): Promise<ArticleDTO> {
-    return ArticlesService.populateArticleQuery(ArticleModel.find({ _id }));
+    const article = await ArticlesService.populateArticleQuery(ArticleModel.findOne({ _id }));
+    if (isEmpty(article)) {
+      throw InternalError.NotFound();
+    }
+    return new ArticleDTO(article);
   }
 
   async create(article: ArticleDTO, user: UserJwtPayload): Promise<ArticleDTO> {
-    if (isEmpty(article)) {
-      throw InternalError.BadRequest("Article data is missing in request body.");
+    if (!isEmpty(article.tags)) {
+      article.tags = await TagsService.validateTags(article.tags);
     }
 
     article.author = user._id;
     return ArticleModel.create(article);
   }
 
-  async deleteOne(_id: string): Promise<void> {
+  async update(_id: string, articlePayload: ArticleDTO, user: UserJwtPayload): Promise<ArticleDTO> {
+    const article = await ArticleModel.findOne({ _id });
+
+    if (isEmpty(article)) {
+      throw InternalError.BadRequest(`Article with id ${_id} not found`);
+    }
+
+    if (String(article.author) !== user._id && !isAdmin(user)) {
+      throw InternalError.Forbidden("Access denied. User can modify only posts created by himself.");
+    }
+
+    if (Array.isArray(articlePayload.tags)) {
+      article.tags = await TagsService.validateTags(articlePayload.tags);
+    }
+
+    article.title = articlePayload.title;
+    article.description = articlePayload.description;
+    await article.save();
+    return new ArticleDTO(article);
+  }
+
+  async deleteOne(_id: string, user: UserJwtPayload): Promise<void> {
+    const article = await ArticleModel.findOne({ _id });
+
+    if (isEmpty(article)) {
+      throw InternalError.BadRequest(`Article with id ${_id} not found`);
+    }
+
+    if (String(article.author) !== user._id && !isAdmin(user)) {
+      throw InternalError.Forbidden("Forbidden. You can delete only your own posts");
+    }
     await ArticleModel.deleteOne({ _id });
     return;
   }
@@ -65,7 +106,7 @@ class ArticlesService {
           likes: user._id
         }
       },
-      { new: true }
+      { new: true, populate: ArticlesService.buildArticlePopulateOptions() }
     );
   }
 
@@ -77,69 +118,19 @@ class ArticlesService {
           likes: user._id
         }
       },
-      { new: true }
+      { new: true, populate: ArticlesService.buildArticlePopulateOptions() }
     );
   }
 
-  async addComment(articleID: string, commentID: string): Promise<ArticleDTO> {
-    return ArticleModel.findByIdAndUpdate(
-      articleID,
+  async addComment(articleID: string, commentID: string): Promise<void> {
+    await ArticleModel.updateOne(
+      { _id: articleID },
       {
         $addToSet: {
           comments: commentID
         }
-      },
-      { new: true }
-    );
-  }
-
-  async addTag(articleID: string, tag: string, user: UserJwtPayload): Promise<ArticleDTO> {
-    const article = await this.getOne(articleID);
-    if (isEmpty(article)) {
-      throw InternalError.NotFound(`Article with id ${articleID} not found`);
-    }
-
-    if (String(article.author) !== user._id && !isAdmin(user)) {
-      throw InternalError.Forbidden("Access denied. User can add tags only to posts created by himself.");
-    }
-
-    const desiredTag = await TagsService.getOne(tag);
-    if (!desiredTag) {
-      throw InternalError.NotFound(`Tag '${tag}' not found`);
-    }
-
-    return ArticleModel.findByIdAndUpdate(
-      articleID,
-      {
-        $addToSet: {
-          tags: tag
-        }
-      },
-      { new: true }
-    );
-  }
-
-  async removeTag(articleID: string, tag: string, user: UserJwtPayload): Promise<void> {
-    const article = await this.getOne(articleID);
-    if (isEmpty(article)) {
-      throw InternalError.NotFound(`Article with id ${articleID} not found`);
-    }
-
-    if (String(article.author) !== user._id && !isAdmin(user)) {
-      throw InternalError.Forbidden(
-        "Access denied. User can remove tags only from posts created by himself."
-      );
-    }
-
-    await ArticleModel.updateOne(
-      { _id: articleID },
-      {
-        $pull: {
-          tags: tag
-        }
       }
     );
-    return;
   }
 }
